+++ client/src/conn/h3.rs
@@ -38,7 +38,7 @@ impl Conn {
 
         // Get an existing pooled connection or establish a new one.
         let quic_conn = match h3
-            .get_or_create_quic_conn(&origin, &host, port, &self.config, &self.server_config)
+            .get_or_create_quic_conn(&origin, &host, port, &self.config, &self.context)
             .await
         {
             Ok(conn) => conn,
@@ -91,15 +91,15 @@ impl Conn {
         }
 
         let transport = self.transport.as_mut().ok_or(Error::Closed)?;
-        let max_buf = self.server_config.http_config().response_buffer_max_len();
+        let max_buf = self.context.http_config().response_buffer_max_len();
         let mut bufwriter = BufWriter::new_with_buffer(
-            Vec::with_capacity(self.server_config.http_config().response_buffer_len()),
+            Vec::with_capacity(self.context.http_config().response_buffer_len()),
             transport,
             max_buf,
         );
 
         let initial_cap = self
-            .server_config
+            .context
             .http_config()
             .request_buffer_initial_len();
         let max_peer_field_section_size = None;
@@ -114,7 +114,7 @@ impl Conn {
             bufwriter.buffer_mut(),
         )?;
 
-        let copy_loops_per_yield = self.server_config.http_config().copy_loops_per_yield();
+        let copy_loops_per_yield = self.context.http_config().copy_loops_per_yield();
 
         if let Some(body) = self.request_body.take() {
             let mut body = body.into_h3();