+++ http/src/conn/h3.rs
@@ -74,15 +74,15 @@ where
         self.finalize_response_headers_h3();
 
         let mut output_buffer =
-            Vec::with_capacity(self.server_config.http_config.response_buffer_len);
+            Vec::with_capacity(self.context.http_config.response_buffer_len);
 
         self.encode_headers_h3(&mut output_buffer)?;
 
-        let loops_per_yield = self.server_config.http_config.copy_loops_per_yield;
+        let loops_per_yield = self.context.http_config.copy_loops_per_yield;
         let max_peer_field_section_size = self.max_peer_field_section_size();
-        let initial_cap = self.server_config.http_config.request_buffer_initial_len;
+        let initial_cap = self.context.http_config.request_buffer_initial_len;
 
-        let max_buf = self.server_config.http_config.response_buffer_max_len;
+        let max_buf = self.context.http_config.response_buffer_max_len;
         let mut bufwriter = BufWriter::new_with_buffer(output_buffer, &mut self.transport, max_buf);
 
         if self.method != Method::Head
@@ -118,7 +118,7 @@ where
         encode_field_section_h3(
             &field_section,
             self.max_peer_field_section_size(),
-            self.server_config.http_config.request_buffer_initial_len,
+            self.context.http_config.request_buffer_initial_len,
             buffer,
         )
     }
@@ -183,14 +183,14 @@ where
         }
 
         let response_headers = h3_connection
-            .server_config()
+            .context()
             .shared_state()
             .get::<Headers>()
             .cloned()
             .unwrap_or_default();
 
         Ok(Conn {
-            server_config: h3_connection.server_config(),
+            context: h3_connection.context(),
             transport,
             request_headers,
             method,